Fix Window function bug
Previously, the Window function would produce proper windows if pixel_size = 2 and would break for other cases.
I figured out that it was because how I was defining nwindows shape. I decided to instead just use the regular
Python list and append instead of calculating the shape of nwindows. It is currently working.
Issues: For some cases, the code ignores the very last window. I am unsure about how to get the indexing to work so
I have left it that way for now.